# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Aquachile",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/shift": {
            "get": {
                "tags": [
                    "Shifts"
                ],
                "summary": "Show all shifts",
                "description": "Returns a list of shifts",
                "operationId": "1f91100716408eac2147c6344b348818",
                "responses": {
                    "200": {
                        "description": "Show all shifts."
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/shift/{id}": {
            "get": {
                "tags": [
                    "Shifts"
                ],
                "summary": "Show shift by ID",
                "description": "Returns an shift",
                "operationId": "7e3303c17cba7dab923a97aaa46a829e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Resource ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show only one shift."
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/shift/active": {
            "get": {
                "tags": [
                    "Shifts"
                ],
                "summary": "Show the active shift",
                "description": "Returns an active shift",
                "operationId": "031004fffa3feed352e97e6e87b66872",
                "responses": {
                    "200": {
                        "description": "Show the active shift."
                    },
                    "404": {
                        "description": "No active shift."
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/shift/open": {
            "post": {
                "tags": [
                    "Shifts"
                ],
                "summary": "Start a new shift",
                "description": "Start a new shift",
                "operationId": "c2256dc0ac609173feda3d50de037641",
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/shift/close": {
            "post": {
                "tags": [
                    "Shifts"
                ],
                "summary": "Close the active shift",
                "description": "Returns a active shift",
                "operationId": "b63d2deb180564633a0247df44b24dea",
                "responses": {
                    "200": {
                        "description": "Close the active shift."
                    },
                    "404": {
                        "description": "No active shift."
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}